Initial app module introduction.
This is based on the empty project generated by Android Studio. It
includes an adaptive Oppia app logo based on the prototype app. It also
includes a shared test that can run both with Robolectric and Espresso.
The initial screen is a basic view with an action bar and a welcome
string that's being verified in two different ways in tests.