Fix #5: GAE Abstraction part 1 - Topic Page handler [Blocked: #85] (#78)
* Initial introduction of the data source module.

High-level notes:
- Data sources are now being referred to as 'DataProviders'
- This commit also fixes the build which was broken in a previous PR
  ('compile' needs to be used for the proto library)
- The data source module is referred to as just 'data'
- This introduces a few style changes to prohibit wildcard imports
- This also normalizes the JDK versions needed across modules, potentially
  fixing some of the repeated reversions the IDE does for a Java SDK
  version setting

Functional differences: the app now properly shows the correct string upon
initial open ("Welcome to Oppia"), and reopening thereafter correctly
shows the updated string ("Welcome back to Oppia"). Rotating the device is
currently broken without Dagger integration.

Overall, this moves all general data provider functionality into the new
module, and introduces a PersistentCacheStore to store proto messages
on-disk, or load them into memory if they are available. Overall, this
commit is introducing a potentially reasonable medium-term solution for
data providers and on-disk proto storage. A long-term design should still
be determined, but this seems like a good place to start.

Tests are currently broken or missing. New TODOs will also be resolved in
a later commit.

* Data module dependencies, retrofit and models

* Removed kapt from build.gradle in data

* Javadoc comments

* Introduce one possible integration with Dagger 2.

This is loosely based on both
https://github.com/tfcporciuncula/dagger-simple-way and
https://dagger.dev/android.html, except this approach is meant to keep
activities and fragments as thin as possible and push as much UI
presentation and service business logic into Dagger-provided objects as
possible. This maximizes Dagger use throughout the codebase, and opens the
potential for future code generation for these thin adapter fragments &
services.

This fixes the rotation bug in the app: rotating upon the initial app open
retains the 'Welcome to Oppia' label since the same user app history
controller is used in both instances of HomeFragment due to it being
singleton scoped and not needing to be recreated.

No tests were updated with these changes.

* Javadoc comments added

* Reolved Javadoc mistakes

* Renaming test cases

* Test cases for TopicService

* Made changes to OppiaGaeClient

* Made changes to OppiaGaeClient

* TopicService test code using Json response

* Revert changes in codeStyles/Project.xml

* Revert changes in codeStyles/Project.xml

* Optimise imports

* Check NetworkInterceptor part

* Fix part of #9: Introduce Retrofit classroom data handler service (#83)

* introduced topic index handler

* Call retrofit data handlers from data module

* Update topic_index_items.xml

* removed extra space

* fixed issues

* Update AndroidManifest.xml

* Removed Ui part. http handlers are included in data module

* done changes in handler and data as per the specs

* Update TopicSummarytDicts.kt

* Update NetworkSettings.kt

* added classroom datahandler

* renamed corresponding classes to classroom

* fixed issues

* Update GaeClassroom.kt

* removed spaces

* Update AndroidManifest.xml

* removed unwanted spaces

* added space

* Resolve typo

* NetworkModule introduced

* Daggerify NetworkInterceptor

* Make retrofit dependency as api

* Javadocs and linking of oppia-web to models

* Changes in Javadoc, naming and test cases code

* Updated TODO comments and NetworkInterceptorTest