Refactor the builder for BindableAdapter. (#409)
This splits the adapter into two: one for single-type recycler views,
and another for multi-type recycler views. The latter has also been
updated to require an enum for typing rather than an integer to ensure
type safety. Per existing usages, we may want to switch this to a KClass
in the future since it seems like we typically convert from class type
to enum, anyway.