Fix #375: Shorten submitted answers (#410)
* Refactor the builder for BindableAdapter.

This splits the adapter into two: one for single-type recycler views,
and another for multi-type recycler views. The latter has also been
updated to require an enum for typing rather than an integer to ensure
type safety. Per existing usages, we may want to switch this to a KClass
in the future since it seems like we typically convert from class type
to enum, anyway.

* Display the original submitted answer text instead of a read-only
version of the interaction with a normalized version of the answer.

* Add test for updated exploration progress controller.

* Remove isReadOnly and previously submitted InteractionObject since
UserAnswer and submitted answer item replaces this mechanism.

* Post-merge fix.