Update all tests to use paused looper with Robolectric.
This is the recommended mode to run all Robolectric tests, and it
interoperates better with TestCoroutineDispatchers.

Note that this commit also includes some functional changes to
AudioPlayerControllerTest. Before, the audio track being played was 1
second long, and 1 second needs to be waited before audio is detected as
playing. These race against each otehr in paused looper mode, so
extending the track duration to 2 seconds fixes the situation. I also
made some minor cleanups elsewhere in the test.