Fix part of #89: Clean up test dispatchers (#1624)
* Introduce test coroutine dispatchers support in Espresso.

This piggybacks off of the solution introduced in #1276 for Robolectric.
That PR allows Robolectric tests in app module to share dependencies
with production components by creating a test application & telling
Robolectric to use it instead of OppiaApplication via a @Config
annotation.

This PR achieves the same thing by using a custom test runner that reads
the same annotation and forces Espresso & instrumentation to use the
test application instead of the default OppiaApplication. This setup
will be easier once #59 is finished since we can specify the application
in per-test manifests that both Robolectric & Espresso will respect.

Note that while this lets the same test coroutine dispatchers work in
both production & test code for Espresso, some additional work was
needed to ensure the coroutines behave correctly. In particular, the
coroutines use a fake system clock in Robolectric that requires explicit
synchronization points in the tests to allow the clock to move forward
& properly coordinate execution between background & main thread
operations. However, in Espresso, since everything is using a real clock
an idling resource is the preferred way to synchronize execution: it
allows the background coroutines to operate in real-time much like they
would in production, and then notify Espresso when completed. The test
dispatchers API is now setup to support both synchronization mechanisms
for both Robolectric & Espresso (the idling resource does nothing on
Robolectric and the force synchronization effectively does nothing on
Espresso).

The first test being demonstrated as now stable is SplashActivityTest
(as part of downstream work in #1397.

* Revert "Fixes #941: Add radar effect in Hints and solution (#1475)"

This reverts commit 41eb10bd0c04596cb18c354da7db7268121a09be.

* Add thorough documentation for new dispatchers.

* Broadly clean up test dispatchers.

TestCoroutineDispatchers have been used in a few heavy Robolectric tests
for a couple months now, without issue. This change broadly migrates all
test dispatchers over to TestCoroutineDispatchers and
TestDispatcherModule, resulting in significant and broad cleanups.

This change also includes broadly removing now-unnecessary uses of
various Kotlin coroutine annotations, and runBlockingTest() since
TestCoroutineDispatchers replaces this functionality & does it more
correctly.

Some things to note:
1. This change also includes moving ProfileTestHelper to the testing
module since it's been changed to rely on TestCoroutineDispatcher (which
improves downstream test stability in the app module).

2. MainThreadExecutor was removed as the test disaptcher in
StateFragmentTest and HomeActivityTest which may cause failures in these
tests. Since neither test is part of the non-flaky test run, this is
considered fine here.

3. Notice that runCurrent() has been moved around or added in a bunch of
places--this is to be expected. TestCoroutineDispatchers always operates
in a paused state, whereas before all test dispatchers were generally
operating in a running state. These additional calls are needed to
synchronize access. In some special cases, an additional coroutine
dispatcher is being used to demonstrate blocking functionality without
interfering with TestCoroutineDispatcher functionality. Also, note that
missing runCurrent() in app module tests that depend on waitFor() can
result in long failures (30 seconds for the timeout), so long-term tests
should opt to avoid using waitFor() since TestCoroutineDispatchers also
has an idling resource for Espresso use cases.

* Update all tests to use paused looper with Robolectric.

This is the recommended mode to run all Robolectric tests, and it
interoperates better with TestCoroutineDispatchers.

Note that this commit also includes some functional changes to
AudioPlayerControllerTest. Before, the audio track being played was 1
second long, and 1 second needs to be waited before audio is detected as
playing. These race against each otehr in paused looper mode, so
extending the track duration to 2 seconds fixes the situation. I also
made some minor cleanups elsewhere in the test.

* Add testing package to CI & ktlint.

* Clean up comments & add additional documentation.

* Fix lint errors.

* Fix broken test after changes to FakeSystemClock.

* Fix linter failures.

* Fix broken BindableAdapterTest post update with develop.

The adapter test requires synchronization when running with a paused
looper mode to make sure state is propagated when desired.

* Clean up new test post-merge.