Add support for checking for prohibited patterns
Add support for checking for prohibited patterns both in file contents & file names. The initial checks being added are:
- Prevent synthetic Kotlin imports
- Prevent activities from being introduced without corresponding presenters
- Prevent fragments from being introduced without corresponding fragments

This script is expected to be heavily expanded in the future to check many other types of patterns. We may want to build some python scripts to simplify the checks, as well, to make it easier to add more. The current bash-heavy versions are a bit hard to maintain.

Also note that this check will not be required initially since there are some existing failures of these checks that first need to be fixed.